+++ tests/Console/ModelsCommand/LaravelCustomCasts/Models/CustomCast.php
@@ -15,6 +15,7 @@ use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Cas
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Casts\CustomCasterWithPrimitiveDocblockReturn;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Casts\CustomCasterWithPrimitiveReturn;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Casts\CustomCasterWithReturnType;
+use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Casts\CustomCasterWithStaticReturnType;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Casts\ExtendedSelfCastingCasterWithStaticDocblockReturn;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Casts\ExtendedSelfCastingCasterWithThisDocblockReturn;
 use Barryvdh\LaravelIdeHelper\Tests\Console\ModelsCommand\LaravelCustomCasts\Casts\InboundAttributeCaster;
@@ -26,6 +27,7 @@ class CustomCast extends Model
 {
     protected $casts = [
         'casted_property_with_return_type' => CustomCasterWithReturnType::class,
+        'casted_property_with_static_return_type' => CustomCasterWithStaticReturnType::class,
         'casted_property_with_return_docblock' => CustomCasterWithDocblockReturn::class,
         'casted_property_with_return_docblock_fqn' => CustomCasterWithDocblockReturnFqn::class,
         'casted_property_with_return_primitive' => CustomCasterWithPrimitiveReturn::class,